requirements.txt
asgiref==3.7.2
Django==5.0
django-nested-admin==4.0.2
python-monkey-business==1.0.0
six==1.16.0
sqlparse==0.4.4
